% Auto-generated: do not edit by hand
\name{dashPannellum}

\alias{dashPannellum}

\title{DashPannellum component}

\description{

}

\usage{
dashPannellum(id=NULL, currentScene=NULL, customControls=NULL,
height=NULL, loaded=NULL, multiRes=NULL, pitch=NULL,
showCenterDot=NULL, tour=NULL, video=NULL, width=NULL,
yaw=NULL)
}

\arguments{
\item{id}{Character. }

\item{currentScene}{Character. }

\item{customControls}{Logical. }

\item{height}{Character. }

\item{loaded}{Logical. }

\item{multiRes}{Named list. }

\item{pitch}{Numeric. }

\item{showCenterDot}{Logical. }

\item{tour}{Named list. }

\item{video}{Lists containing elements 'sources', 'poster'.
those elements have the following types:
  - sources (required): . sources has the following type: list of lists containing elements 'src', 'type'.
those elements have the following types:
  - src (character; required)
  - type (character; required)s
  - poster (character; optional). }

\item{width}{Character. }

\item{yaw}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

